Handle playback issues with tabbed docked widgets with Qt 4.8.
With Qt 4.8, we have started seeing change in the order of the
tab-widgets created for docked panels. That results in the names not
aligning to the ones used while recording the test. In that case, we
can attempt to locate a sibling tab-bar that has a pane with the expected
name. If none is found, then alone do we give up.