Add public function to allow disabling the playing of events on blocking.
This is useful if a played-back event happens to run an event loop, but
isn't a modal dialog, and the player knows the event should complete before
anything else runs. For example, playing back an event that starts a
blocking event loop waiting for a background thread to finish its work.