Add guard for context menu in some view event players
The pqAbstractItemViewEventPlayer, pqTreeViewEventPlayer and pqAbstractItemViewEventPlayerBase all try to handle an object command for the object's parent if the object itself is not what the players expects. This will not work in case of a context menu for these views, because the context menu's parent is the view, and in some cases, the view and the context menu commands are same, for example, "activate", so the view player will try to handle the command for context menu, which apparently will fail. The fix is not to look at the parent if it is a QMenu.