Enable use of continuous flush
Continous flush options was available, however not accessible.
This commit add a checkbox to control this options,
Modify slughtly the signal api so it is standardized and actually use the signal
in the only provided observer.
When true, continuous flush ensure that events are written on file
as sson as they are recorded.