Fixes for Qt 5 / OsX test playback.
With Qt 5, on OsX all file dialog (and other modal dialog) tests would
fail. The problem was that the modal dialog's event loop would not quite
even after `QDialog::accept()` was called. This happened when
`QApplication::processEvents` was called by the testing framework when
processing a test item while the modal dialog was open. Skipping the
calling of `QApplication::processEvents` overcomes that issue.