Block recording commands using a "BlockRecordCommands" dynamic property
The motivation for this is to avoid recording commands that are
triggered programmatically. Many translators connect to signals that are
fired both when a user interacts with the widget and when methods are
called by the programmer.  For example the
pqAbstractItemViewEventTranslatorBase connects to the currentChanged
signal.

Recording programmatic events is problematic because the previous
command in the script has already triggered the recorded programmatic
changes.  This unneessarily complicates the script and also causes the
programmatic commands to be executed twice.  I also encountered a case
where programmatic events were recorded *before* the user event which
created an invalid script.